general: suppress some CodeQL warnings CODABLOCKF: tidy some comments

+++ backend/dmatrix.c
@@ -1904,7 +1904,7 @@ static int dm_ecc200(struct zint_symbol *symbol, struct zint_seg segs[], const i
         const int NC = W - 2 * (W / FW);
         const int NR = H - 2 * (H / FH);
         int x, y, *places;
-        if (!(places = (int *) calloc(NC * NR, sizeof(int)))) {
+        if (!(places = (int *) calloc((size_t) NC * (size_t) NR, sizeof(int)))) {
             return errtxt(ZINT_ERROR_MEMORY, symbol, 718, "Insufficient memory for placement array");
         }
         dm_placement(places, NR, NC);
